backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
google-generativeai==0.8.3


